fix: split elero.cpp into 3 files to prevent OOM during compilation
The 2032-line elero.cpp caused the xtensa compiler (cc1plus) to be
killed by the OOM killer on memory-constrained build environments.

Split into:
- elero.cpp (~558 lines): core lifecycle, HAL, setup, loop, radio task
- elero_cc1101.cpp (~589 lines): CC1101 register access, TX/RX state machines, watchdog
- elero_protocol.cpp (~786 lines): protocol crypto, message dispatch, runtime blinds

ESPHome auto-discovers .cpp files in component directories, so no
build system changes are needed.
